Atmospheric charger
Ported atmospheric charger
*Simplified model- now a single cube
*Instead of the large rendered antenna, it now requires 3 iron bar blocks in a column on top. Block type for antenna controlled via tag
*Redid animations. Untested, because I still can't compile, but they should be an improvement.

+++ src/main/java/com/Da_Technomancer/crossroads/API/alchemy/AtmosChargeSavedData.java
@@ -11,7 +11,7 @@ public class AtmosChargeSavedData extends WorldSavedData{
 
 	public static final String ID = Crossroads.MODID + "_atmos";
 
-	public static long getCapacity(){
+	public static int getCapacity(){
 		return CRConfig.atmosCap.get();
 	}
